requirements.txt
asgi-prometheus==1.1.0
fastapi==0.92.0
jinja2==3.1.2
Pillow==9.4.0
pydantic[dotenv]
uvicorn
